frameworks/chargerd: support capacity threshold to control chargering
When the battery level is within a certain range, it does not charge. If
the battery level drops below the left boundary of the range, charging
begins and continues until it reaches the right boundary, at which point
charging stops. When the battery level is within this range and a
charger is plugged in, it will not charge.

Signed-off-by: dongjiuzhu1 <dongjiuzhu1@xiaomi.com>